requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.25.0

# Vision-Language Models
lmdeploy>=0.5.0
qwen-vl-utils>=0.0.2
peft>=0.7.0

# Fine-tuning
llamafactory>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Image downloading
requests>=2.31.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=0.5.1
pyyaml>=6.0
nest-asyncio>=1.5.0
click>=8.0.0

# Development
flake8>=6.0.0
Sphinx>=7.0.0

# Local package
-e .
